// src/adapters/llm/types.ts
export interface OllamaGenerateResponse {
  model: string;
  response: string;
  done: boolean;
  context: number[];
  total_duration: number;
  load_duration: number;
  prompt_eval_count: number;
  prompt_eval_duration: number;
  eval_count: number;
  eval_duration: number;
}

export interface OllamaGenerateStreamChunk {
  model: string;
  response: string;
  done: boolean;
}